+++ packages/vira/src/Vira/CI/Pipeline.hs
@@ -3,35 +3,22 @@
 
 module Vira.CI.Pipeline (runPipeline, defaultPipeline, PipelineError (..)) where
 
-import Attic qualified
-import Attic.Config (lookupEndpointWithToken)
-import Attic.Types (AtticServer (..), AtticServerEndpoint)
-import Attic.Url qualified
-import Control.Monad.Writer.Strict (MonadWriter (tell), WriterT (..))
 import Effectful (Eff, IOE, (:>))
 import Effectful.Error.Static (Error, runErrorNoCallStack, throwError)
 import Effectful.Git qualified as Git
-import Effectful.Process (CreateProcess)
-import GH.Signoff qualified as Signoff
 import Language.Haskell.Interpreter (InterpreterError (..))
 import Shower qualified
 import System.Directory (doesFileExist)
 import System.Exit (ExitCode (ExitSuccess))
 import System.FilePath ((</>))
-import System.Info qualified as SysInfo
 import Vira.CI.Configuration qualified as Configuration
 import Vira.CI.Environment (ViraEnvironment (..))
 import Vira.CI.Environment qualified as Env
 import Vira.CI.Error
 import Vira.CI.Pipeline.Type
-import Vira.Lib.Omnix qualified as Omnix
-import Vira.Lib.Process (alwaysUnderPath)
+import Vira.CI.Processes (pipelineProcesses)
 import Vira.State.Type (cloneUrl, headCommit)
 import Vira.Supervisor.Task qualified as Task
-import Vira.Tool.Core (ToolError (..))
-import Vira.Tool.Tools.Attic qualified as AtticTool
-import Vira.Tool.Type.ToolData (status)
-import Vira.Tool.Type.Tools (attic)
 
 -- | Run `ViraPipeline` for the given `ViraEnvironment`
 runPipeline ::
@@ -60,71 +47,6 @@ runPipeline env = do
               Task.runProcesses pipelineProcs <&> first PipelineTerminated
     Right exitCode -> do
       pure $ Right exitCode
-
-{- | Get all the processes to run as part of this Pipeline
-
-TODO: To be rewritten during https://github.com/juspay/vira/issues/6
--}
-pipelineProcesses :: ViraEnvironment -> ViraPipeline -> Either PipelineError (NonEmpty CreateProcess)
-pipelineProcesses env pipeline = do
-  (_, postBuildProcs) <- runWriterT $ do
-    tell <=< lift $ cacheProcs env pipeline.cache
-    tell $ signoffProcs pipeline.signoff
-  let procs = buildProc pipeline.build :| postBuildProcs
-  pure $ procs <&> alwaysUnderPath Env.projectDirName
-
-buildProc :: BuildStage -> CreateProcess
-buildProc stage =
-  Omnix.omnixCiProcess (overrideInputsToArgs stage.overrideInputs)
-  where
-    -- Convert override inputs to command line arguments
-    overrideInputsToArgs :: [(Text, Text)] -> [String]
-    overrideInputsToArgs =
-      concatMap (\(key, value) -> ["--override-input", toString key, toString value])
-
-cacheProcs :: ViraEnvironment -> CacheStage -> Either PipelineError [CreateProcess]
-cacheProcs env stage =
-  go env.tools.attic stage.url
-  where
-    go _ Nothing = pure []
-    go attic (Just urlText) = do
-      -- Parse cache URL once
-      (serverEndpoint, cacheName) <-
-        Attic.Url.parseCacheUrl urlText
-          & first (parseErrorToPipelineError urlText)
-
-      -- Get attic config and create push process
-      pushProc <- first (atticErrorToPipelineError urlText serverEndpoint) $ do
-        atticConfig <- attic.status
-        -- Get server name for endpoint (only if it has a token)
-        serverName <-
-          lookupEndpointWithToken atticConfig serverEndpoint
-            & maybeToRight (AtticTool.MissingEndpoint serverEndpoint)
-        -- Create the push process (token already validated by lookupEndpointWithToken)
-        pure $ Attic.atticPushProcess (AtticServer serverName serverEndpoint) cacheName "result"
-      pure $ one pushProc
-
-    parseErrorToPipelineError :: Text -> Attic.Url.ParseError -> PipelineError
-    parseErrorToPipelineError url err =
-      PipelineConfigurationError $
-        MalformedConfig $
-          "Invalid cache URL '" <> url <> "': " <> show err
-
-    atticErrorToPipelineError :: Text -> AtticServerEndpoint -> AtticTool.ConfigError -> PipelineError
-    atticErrorToPipelineError url _endpoint err =
-      let suggestion = AtticTool.configErrorToSuggestion err
-          suggestionText =
-            "\n\nSuggestion: Run the following in your terminal\n\n"
-              <> show @Text suggestion
-          msg = "Attic configuration error for cache URL '" <> url <> "': " <> show err <> suggestionText
-       in PipelineToolError $ ToolError msg
-
-signoffProcs :: SignoffStage -> [CreateProcess]
-signoffProcs stage =
-  [Signoff.create Signoff.Force statusTitle | stage.enable]
-  where
-    nixSystem = SysInfo.arch <> "-" <> SysInfo.os
-    statusTitle = "vira/" <> nixSystem <> "/ci"
 
 {- | Load pipeline configuration for the given project environment.
